requirements.txt
anyio==3.6.1
autopep8==1.6.0
click==8.1.3
colorama==0.4.5
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
psycopg2==2.9.3
pycodestyle==2.8.0
pydantic==1.9.1
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.2
sse-starlette==1.1.6
colorlog==6.6.0
python-dotenv==0.20.0